2018.12.28 17:40

+++ smartGym-backEnd/smartGym-backEnd-dao/src/main/java/cn/smartGym/mapper/SmartgymPlayersMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface SmartgymPlayersMapper {
-    long countByExample(SmartgymPlayersExample example);
+    int countByExample(SmartgymPlayersExample example);
 
     int deleteByExample(SmartgymPlayersExample example);
